# openapi/api.yaml
openapi: 3.0.3
info:
  title: Shared Study Logger API
  version: 3.0.0
servers:
  - url: /
paths:
...
  /api/auth/login:
    post:
      operationId: postAuthLogin
      summary: ログイン
      parameters:
        - $ref: "#/components/parameters/XClientApiVersion"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: ログイン成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        default:
          $ref: "#/components/responses/Error"
  /api/auth/logout:
    post:
      operationId: postAuthLogout
      summary: ログアウト
      security:
        - session: []
      parameters:
        - $ref: "#/components/parameters/XClientApiVersion"
      responses:
        "200":
          description: ログアウト成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
        default:
          $ref: "#/components/responses/Error"
  /api/auth/me:
    get:
      operationId: getAuthMe
      summary: ログイン中ユーザー情報取得
      security:
        - session: []
      parameters:
        - $ref: "#/components/parameters/XClientApiVersion"
      responses:
        "200":
          description: ユーザー情報
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        default:
          $ref: "#/components/responses/Error"
    patch:
      operationId: patchAuthMe
      summary: プロフィール更新
      security:
        - session: []
      parameters:
        - $ref: "#/components/parameters/XClientApiVersion"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
      responses:
        "200":
          description: 更新後ユーザー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        default:
          $ref: "#/components/responses/Error"
  /api/auth/password:
    post:
      operationId: postAuthPassword
      summary: パスワード変更
      security:
        - session: []
      parameters:
        - $ref: "#/components/parameters/XClientApiVersion"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordRequest"
      responses:
        "200":
          description: 変更成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
        default:
          $ref: "#/components/responses/Error"
  /api/users/{userId}:
    get:
      operationId: getUserById
      summary: 公開プロフィール取得
      security:
        - session: []
      parameters:
        - $ref: "#/components/parameters/XClientApiVersion"
        - $ref: "#/components/parameters/userId"
      responses:
        "200":
          description: 公開プロフィール
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicUserResponse"
        default:
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: session
  parameters:
    XClientApiVersion:
      name: X-Client-Api-Version
      in: header
      required: true
      description: クライアント API 版。info.version および CLIENT_API_VERSION と同じ major.minor.patch。
      schema:
        $ref: "#/components/schemas/ClientApiVersion"
...
    userId:
      name: userId
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/ResourceId"
...
  responses:
    Error:
      description: エラー
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
    ClientApiVersion:
      type: string
      description: major.minor.patch。CLIENT_API_VERSION_PATTERN_SOURCE と同じ。
      pattern: '^(0|[1-9]\d*)\.(0|[1-9]\d*)\.(0|[1-9]\d*)$'
    ResourceId:
      type: string
      format: uuid
    Timestamp:
      type: string
      format: date-time
...
    AvatarKey:
      type: string
      enum:
        - avoidy
        - lavender
        - fever
        - born
        - innocence
        - stolen
    UserRole:
      type: string
      enum:
        - ADMIN
        - USER
...
    OkResponse:
      type: object
      required:
        - ok
      properties:
        ok:
          type: boolean
          enum:
            - true
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 1
    User:
      type: object
      required:
        - id
        - email
        - displayName
        - role
        - bio
        - avatarKey
        - createdAt
      properties:
        id:
          $ref: "#/components/schemas/ResourceId"
        email:
          type: string
          format: email
        displayName:
          type: string
        role:
          $ref: "#/components/schemas/UserRole"
        bio:
          type: string
          nullable: true
        avatarKey:
          nullable: true
          allOf:
            - $ref: "#/components/schemas/AvatarKey"
        createdAt:
          $ref: "#/components/schemas/Timestamp"
    UserResponse:
      type: object
      required:
        - user
      properties:
        user:
          $ref: "#/components/schemas/User"
...
    PublicUser:
      type: object
      required:
        - id
        - displayName
        - bio
        - avatarKey
        - createdAt
      properties:
        id:
          $ref: "#/components/schemas/ResourceId"
        displayName:
          type: string
        bio:
          type: string
          nullable: true
        avatarKey:
          nullable: true
          allOf:
            - $ref: "#/components/schemas/AvatarKey"
        createdAt:
          $ref: "#/components/schemas/Timestamp"
    PublicUserResponse:
      type: object
      required:
        - user
      properties:
        user:
          $ref: "#/components/schemas/PublicUser"
    UpdateProfileRequest:
      type: object
      required:
        - displayName
        - bio
        - avatarKey
      properties:
        displayName:
          type: string
          minLength: 1
          maxLength: 50
        bio:
          type: string
          maxLength: 500
          nullable: true
        avatarKey:
          nullable: true
          allOf:
            - $ref: "#/components/schemas/AvatarKey"
    ChangePasswordRequest:
      type: object
      required:
        - currentPassword
        - newPassword
      properties:
        currentPassword:
          type: string
          minLength: 1
        newPassword:
          type: string
          minLength: 8
          maxLength: 128
...